pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pathoflow"
version = "0.1.0"
description = "High-performance WSI inference engine"
requires-python = ">=3.9"
dependencies = [
    "openslide-python",
    "numpy",
    "opencv-python-headless",
    "pydantic",
    "typer",
    "torch",
    "torchvision",
    "matplotlib",
    "pytest"
]

[project.scripts]
pathoflow = "pathoflow.cli:app"

[tool.setuptools.packages.find]
where = ["src"]
